// src/app/model/filterTag/filter.model.ts
import { FilterTag } from "./filterTag";

export class FilterModel{
    activity?: FilterTag[];
    characteristics?: FilterTag[];
    crops?: FilterTag[];
    composition?: FilterTag[];
    phenologicalPhase?: FilterTag[];
    formulation?: FilterTag[];
    activeIngredient?: FilterTag[];
    // cropProtection?: FilterTag[];
}